register fix

+++ internal/domain/model/user.go
@@ -2,7 +2,7 @@ package model
 
 type User struct {
 	ID           int64
-	AppID        int
+	AppID        int64
 	Email        string
 	PasswordHash []byte
 	IsAdmin      bool
