+++ internal/grpc/auth/requests/login_request.go
@@ -3,5 +3,5 @@ package requests
 type LoginRequest struct {
 	Email    string `validate:"required,email"`
 	Password string `validate:"required"`
-	AppID    int32  `validate:"required"`
+	AppID    int64  `validate:"required"`
 }
